+++ middleware.ts
@@ -10,7 +10,6 @@ const allowedOrigins = process.env.NODE_ENV === 'production'
 export function middleware(request: Request) {
 
     const origin = request.headers.get("origin")
-    console.log(origin)
     if (origin && !allowedOrigins.includes(origin)) {
         return new NextResponse(null, {
             status: 400,
@@ -20,11 +19,5 @@ export function middleware(request: Request) {
             }
         })
     }
-
-    console.log(request.method)
-    console.log(request.url)
-
-
-
     return NextResponse.next()
 }
