+++ pages/api/users/getUserInformations.ts
@@ -18,3 +18,21 @@ export async function handleInfosUser(session: any) {
         console.error('Error fetching data:', error);
     }
 }
+
+export async function GetAllStaff(session: any, establishmentId: string) {
+    try {
+        const response = await fetch(process.env.SERVER_API + `/users?establishmentId=${establishmentId}`, {
+            headers: {
+                Authorization: `Bearer ${session.user.tokens.accessToken}`,
+            },
+        });
+        const data = await response.json();
+        console.log("allUsers",data)
+        if (response.status === 200) {
+            console.log(data);
+        }
+        return data;
+    } catch (error) {
+        console.error('Error fetching data:', error);
+    }
+}
