+++ types/IActivity.ts
@@ -1,0 +1,11 @@
+export interface IActivity {
+  color: string;
+  description: string;
+  duration: number;
+  employees: string[];
+  establishment: string;
+  imageUrl: string;
+  price: number;
+  title: string;
+  _id: string;
+}
