+++ types/IDogs.ts
@@ -1,0 +1,10 @@
+export interface IDogs {
+  nationalId: string,
+  name: string,
+  imageUrl: string,
+  gender: string,
+  breed: string,
+  birthDate: string,
+  weight: number,
+  height: number,
+}
