+++ types/IUser.ts
@@ -1,0 +1,13 @@
+export interface IUser {
+  activities: [],
+  _id: string,
+  lastname: string,
+  firstname: string,
+
+  avatarUrl: string,
+  role: string,
+  emailAddress: string,
+  password: string,
+  registeredAt: string,
+  lastConnectionAt: string,
+}
